new changes version 1.0

+++ flickrgallery/src/com/flickrgallery/activity/MainActivity.java
@@ -2,6 +2,7 @@ package com.flickrgallery.activity;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.HashSet;
 
 import android.app.Activity;
 import android.content.Context;
@@ -15,6 +16,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
+import android.widget.AdapterView.OnItemSelectedListener;
 import android.widget.BaseAdapter;
 import android.widget.GridView;
 import android.widget.ImageView;
@@ -22,6 +24,7 @@ import android.widget.ProgressBar;
 import android.widget.SearchView;
 
 import com.flickrgallery.data.Object;
+import com.flickrgallery.data.Photo;
 import com.flickrgallery.observer.Observer;
 import com.flickrgallery.request.GetList;
 import com.flickrgallery.request.GetPhotos;
@@ -39,6 +42,8 @@ public class MainActivity extends Activity implements Observer {
 	private Object result = null;
 	private ImageAdapter imageAdapter;
 	private ProgressBar progressBar;
+	private HashSet<String> filePaths = new HashSet<String>();
+	private File file;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -47,25 +52,57 @@ public class MainActivity extends Activity implements Observer {
 
 		gridView = (GridView) findViewById(R.id.gridView1);
 		progressBar = (ProgressBar) findViewById(R.id.progressBar1);
+		progressBar.setVisibility(View.GONE);
 
 		String Url = Util.LIST_URL + Util.METHOD_GET_PHOTOS + "&"
 				+ Util.API_KEY + "&format=json&nojsoncallback=1";
 		System.out.println("Url:" + Url);
 
-		//TODO: work on file save method ,service mechanism to refresh gallery
+		// TODO: work on file save method ,service mechanism to refresh gallery
 		File file = new File(Util.DIR_PATH);
-		for (File file2 : file.listFiles()) {
-			file2.delete();
-		}
+		if (file.exists())
+			for (File file2 : file.listFiles()) {
+				file2.delete();
+			}
 		File file1 = new File(Util.DIR_PATH_FULL_IMAGE);
-		for (File file2 : file1.listFiles()) {
-			file2.delete();
-		}
-		
-		
+		if (file1.exists())
+			for (File file2 : file1.listFiles()) {
+				file2.delete();
+			}
+
 		GetList getList = new GetList();
 		getList.registerObserver(this);
 		getList.execute(Url);
+
+		// TODO: gridview
+
+		int displayWidth = getResources().getDisplayMetrics().widthPixels;
+
+		int imageWidth = displayWidth / 3;
+		gridView.setColumnWidth(imageWidth);
+		gridView.setStretchMode(GridView.NO_STRETCH);
+		imageAdapter = new ImageAdapter(this);
+		gridView.setAdapter(imageAdapter);
+
+		gridView.setOnItemClickListener(new OnItemClickListener() {
+
+			@Override
+			public void onItemClick(AdapterView<?> parent, View view,
+					int position, long id) {
+
+//				 Photo photo = (Photo) parent.getItemAtPosition(position);
+				
+				 Photo photo = (Photo) view.getTag();
+				 Intent intent = new Intent(MainActivity.this,
+				 ImageActivity.class);
+				 intent.putExtra("id", photo.id);
+				 intent.putExtra("farm", photo.farm);
+				 intent.putExtra("server", photo.server);
+				 intent.putExtra("secret", photo.secret);
+				 intent.putExtra("title", photo.title);
+				 startActivity(intent);
+			}
+		});
 	}
 
 	@Override
@@ -90,66 +127,44 @@ public class MainActivity extends Activity implements Observer {
 			GetPhotos getPhotos = new GetPhotos();
 			getPhotos.registerObserver(this);
 			getPhotos.execute(result);
-		} else if (string.equalsIgnoreCase("DOWNLOAD_COMPLETE")) {
-			progressBar.setVisibility(View.GONE);
-			int displayWidth = getResources().getDisplayMetrics().widthPixels;
-
-			int imageWidth = displayWidth / 3;
-			gridView.setColumnWidth(imageWidth);
-			gridView.setStretchMode(GridView.NO_STRETCH);
-			imageAdapter = new ImageAdapter(this);
-			gridView.setAdapter(imageAdapter);
-
-			gridView.setOnItemClickListener(new OnItemClickListener() {
-
-				@Override
-				public void onItemClick(AdapterView<?> parent, View view,
-						int position, long id) {
-					Intent intent = new Intent(MainActivity.this,
-							ImageActivity.class);
-					intent.putExtra("id",
-							result.photosResult.photosList.get(position).id);
-					intent.putExtra("farm",
-							result.photosResult.photosList.get(position).farm);
-					intent.putExtra("server",
-							result.photosResult.photosList.get(position).server);
-					intent.putExtra("secret",
-							result.photosResult.photosList.get(position).secret);
-					intent.putExtra("title",
-							result.photosResult.photosList.get(position).title);
-					startActivity(intent);
-				}
-			});
-		}
-	}
-
-	public class ImageAdapter extends BaseAdapter {
-
-		public Context context;
-		private ArrayList<String> filePaths = new ArrayList<String>();
-		private File file;
-
-		public ImageAdapter(Context context) {
-			this.context = context;
+		} else if (string.equals("DOWNLOAD_COMPLETE")) {
+			System.out.println("Downloaded imagedd");
+			// progressBar.setVisibility(View.GONE);
 			file = new File(Util.DIR_PATH);
 			System.out
 					.println("============length::" + file.listFiles().length);
 			for (File files : file.listFiles()) {
 				filePaths.add(files.getAbsolutePath());
 			}
+
+			this.runOnUiThread(new Runnable() {
+
+				@Override
+				public void run() {
+					imageAdapter.notifyDataSetChanged();
+				}
+			});
+
+		}
+	}
+
+	public class ImageAdapter extends BaseAdapter {
+
+		public Context context;
+
+		public ImageAdapter(Context context) {
+			this.context = context;
 		}
 
 		@Override
 		public int getCount() {
-			int length = 0;
-			if (file.exists())
-				length = file.list().length;
+			int length = 100;
 			return length;
 		}
 
 		@Override
-		public java.lang.Object getItem(int arg0) {
-			return null;
+		public java.lang.Object getItem(int position) {
+			return result.photosResult.photosList.get(position);
 		}
 
 		@Override
@@ -159,21 +174,41 @@ public class MainActivity extends Activity implements Observer {
 
 		@Override
 		public View getView(int position, View view, ViewGroup parent) {
-			ImageView imageView;
-			Bitmap bitmap = BitmapFactory.decodeFile(filePaths.get(position));
-			if (view == null) {
+			ImageView imageView = null;
+			
+			if (view != null) {
+				imageView = (ImageView) view;
+				imageView.setTag(null);
+			} else {
 				imageView = new ImageView(context);
+				imageView.setImageDrawable(getResources().getDrawable(
+						R.drawable.ic_launcher));
 				int diplaySize = getResources().getDisplayMetrics().widthPixels;
 
 				int imageWidth = diplaySize / 3;
 				imageView.setLayoutParams(new GridView.LayoutParams(imageWidth,
 						imageWidth));
 				imageView.setScaleType(ImageView.ScaleType.FIT_XY);
-			} else {
-				imageView = (ImageView) view;
+				imageView.setTag(null);
 			}
-			imageView.setImageBitmap(bitmap);
+			if (filePaths != null && filePaths.size() > position) {
+				Bitmap bitmap = BitmapFactory
+						.decodeFile(filePaths.toArray()[position].toString());
+				imageView.setTag(null);
+				String [] str = filePaths.toArray()[position].toString().split("/");
+				imageView.setTag(getPhoto(str[str.length-1],result.photosResult.photosList));
+				imageView.setImageBitmap(bitmap);
+			}
 			return imageView;
 		}
+	}
+	
+	public Photo getPhoto(String  id, ArrayList<Photo> arrayList) {
+		for (Photo photo : arrayList) {
+			if (photo.id.equalsIgnoreCase(id)) {
+				return photo;
+			}
+		}
+		return null;
 	}
 }
